requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap4==2.3.1
django-libs==2.0.3
django-login-required-middleware==0.6.1
django-paypal==1.1
django-paypal-express-checkout==1.9.1
idna==2.10
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
Pillow==8.1.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
template==0.6.6
toml==0.10.2
urllib3==1.26.4
